chore: update backend plugin export to use bundler

rh-pre-commit.version: 2.4.0
rh-pre-commit.check-secrets: ENABLED
